+++ src/service/QuestionLogic.java
@@ -1,10 +1,11 @@
-package model;
+package service;
 
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.Random;
 
 import dao.MeganatorDAO;
+import model.Question;
 
 public class QuestionLogic {
 	private final int QUEST_AMOUNT = 10;
