+++ src/service/ResultLogic.java
@@ -1,10 +1,12 @@
-package model;
+package service;
 
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
 import dao.MeganatorDAO;
+import model.Question;
+import model.Result;
 
 public class ResultLogic {
 	/**
